refactor: move inline test modules into sibling files
Files were running to two thousand lines, and roughly half of that was a
trailing `#[cfg(test)] mod tests` block. A reader looking for the code had to
scroll past the tests, and a reader looking for a test had to know it was at the
bottom of the module that happens to contain the thing under test.

Each block moves to a sibling `tests.rs` — `src/graph.rs` gets `src/graph/tests.rs`,
`src/lib.rs` gets `src/tests.rs` — which is the layout `crates/uf_lint` already
used. Nothing else changes: no test renamed, none added, none removed.

    uf_pm/src/detect.rs     2204 -> 1214
    uf_std/src/lib.rs       1607 -> 1422
    uf_rsc/src/graph.rs     1567 ->  986
    uf_rsc/src/action.rs    1066 ->  577
    uf_rsc/src/directive.rs 1014 ->  600
    uf_bundle/src/size.rs    730 ->  452

Crates whose files are still over four hundred lines of source need a real
split by topic rather than this mechanical one; that is the next change.
`uf_cli`, `uf_config`, `uf_fmt`, and `uf_lint` are left alone here because they
have work in flight.
